+++ particula/next/condensation.py
@@ -41,7 +41,7 @@ import numpy as np
 
 # particula imports
 from particula.next.particle import Particle
-from particula.next.gas import GasSpecies
+from particula.next.gas.species import GasSpecies
 from particula.constants import GAS_CONSTANT
 from particula.util.knudsen_number import calculate_knudsen_number
 from particula.util.mean_free_path import molecule_mean_free_path
