+++ particula/next/gas/tests/process_test.py
@@ -7,11 +7,11 @@ import pytest
 import numpy as np
 
 from particula.next.particle import Particle, particle_strategy_factory
-from particula.next.gas.gas_vapor_pressure import ConstantVaporPressureStrategy
-from particula.next.gas.gas_species import GasSpeciesBuilder
-from particula.next.gas.gas_collection import GasBuilder
+from particula.next.gas.vapor_pressure import ConstantVaporPressureStrategy
+from particula.next.gas.species import GasSpeciesBuilder
+from particula.next.gas.atmosphere import AtmosphereBuilder
 from particula.next.aerosol import Aerosol
-from particula.next.gas.gas_process import adiabatic_pressure_change, \
+from particula.next.gas.process import adiabatic_pressure_change, \
     AdiabaticPressureChange
 from particula.next.particle_activity import MassIdealActivity
 from particula.next.surface import surface_strategy_factory
@@ -55,7 +55,7 @@ def sample_gas():
                    .build())
 
     return (
-        GasBuilder()
+        AtmosphereBuilder()
         .temperature(298.15)
         .total_pressure(101325)
         .add_species(gas_species)
