+++ particula/next/tests/aerosol_test.py
@@ -8,9 +8,9 @@
 import pytest
 import numpy as np
 from particula.next.aerosol import Aerosol
-from particula.next.gas import GasBuilder
-from particula.next.gas.gas_species import GasSpeciesBuilder, GasSpecies
-from particula.next.gas.gas_vapor_pressure import ConstantVaporPressureStrategy
+from particula.next.gas.atmosphere import AtmosphereBuilder
+from particula.next.gas.species import GasSpeciesBuilder, GasSpecies
+from particula.next.gas.vapor_pressure import ConstantVaporPressureStrategy
 from particula.next.particle import Particle, particle_strategy_factory
 from particula.next.particle_activity import MassIdealActivity
 from particula.next.surface import surface_strategy_factory
@@ -38,7 +38,7 @@ def sample_gas():
                    .build())
 
     return (
-        GasBuilder()
+        AtmosphereBuilder()
         .temperature(298.15)
         .total_pressure(101325)
         .add_species(gas_species)
